fix: apply all change
