+++ src/components/TodosList.js
@@ -3,16 +3,13 @@ import { TodosListContext } from "../contexts";
 import Todo from "./Todo";
 
 export default function TodosList() {
-  const { data, setData } = useContext(TodosListContext);
-  function deleteTodo(id) {
-    let newValue = data.filter((item) => item.id !== id);
-    setData(newValue);
-  }
+  const { data } = useContext(TodosListContext);
+
   return (
     <div>
-      {data?.length === 0 && <p className="text-2xl">No Todos</p>}
-      {data?.map((item, i) => {
-        return <Todo key={item.id} {...item} deleteTodo={deleteTodo} />;
+      {data.length === 0 && <p className="text-2xl">No Todos</p>}
+      {data.map((item) => {
+        return <Todo key={item.id} {...item} />;
       })}
     </div>
   );
